Struct cmaes::options::CMAESOptions

pub struct CMAESOptions {
    pub end_conditions: Vec<CMAESEndConditions>,
    pub dimension: usize,
    pub threads: usize,
}

A container for end conditions, problem dimension, and thread count.

Fields

end_conditions

dimension

threads

Methods

impl CMAESOptions

fn default(dimension: usize) -> CMAESOptions

Returns a set of default options with the specified dimension (number of variables to optimize).

fn custom(dimension: usize) -> CMAESOptions

Returns a set of options with no end conditions.

fn threads(self, threads: usize) -> CMAESOptions

Sets the number of threads to use in the algorithm.

fn stable_generations(self, fitness: f64, generations: usize) -> CMAESOptions

Sets the stable generation count. The algorithm terminates if the specified number of generations pass where the change in best fitness is under the specified amount.

fn fitness_threshold(self, fitness: f64) -> CMAESOptions

Sets the minimum fitness. The algorithm terminates if the best fitness is under the threshold.

fn max_generations(self, generations: usize) -> CMAESOptions

Sets the maximum generation count. The algorithm terminates after the specified number of generations.

fn max_evaluations(self, evaluations: usize) -> CMAESOptions

Sets the maximum evaluation count. The algorithm terminates after the specified number of fitness function calls.